+++ src/nswcaselaw/nswcaselaw.py
@@ -18,13 +18,15 @@ import argparse
 import csv
 import json
 import logging
+import re
 import sys
+import time
 from pathlib import Path
 
 from nswcaselaw import __version__
-from nswcaselaw.constants import COURTS
+from nswcaselaw.constants import CASELAW_BASE_URL, COURTS
 from nswcaselaw.decision import SCRAPER_WARNING, Decision
-from nswcaselaw.search import Search
+from nswcaselaw.search import DEFAULT_PAUSE, Search
 
 __author__ = "Mike Lynch"
 __copyright__ = "The University of Sydney"
@@ -92,6 +94,11 @@ def parse_args(args):
     parser.add_argument("--legislationCited", type=str)
     parser.add_argument("--casesCited", type=str)
     parser.add_argument(
+        "--pause",
+        type=int,
+        help=f"Seconds to wait between requests, default {DEFAULT_PAUSE}",
+    )
+    parser.add_argument(
         "--courts",
         type=int,
         nargs="+",
@@ -104,6 +111,9 @@ def parse_args(args):
         nargs="+",
         default=[],
         help="Select one or more by index number (see --list tribunals)",
+    )
+    parser.add_argument(
+        "--uris", type=Path, default=None, help="CSV file with caselaw URIs"
     )
     parser.add_argument(
         "--output",
@@ -183,6 +193,7 @@ def run_query(args: argparse.Namespace):
         casesCited=args.casesCited,
         courts=args.courts,
         tribunals=args.tribunals,
+        pause=args.pause,
     )
     n = 0
     with output_stream(args.output) as fh:
@@ -192,11 +203,34 @@ def run_query(args: argparse.Namespace):
                 # only print the header if there's at least one result
                 csvout.writerow(decision.header)
             if args.download:
+                time.sleep(args.pause)
                 download_decision(decision, args)
             csvout.writerow(decision.row)
             n += 1
             if args.limit and n >= args.limit:
                 break
+
+
+def download_uris(args: argparse.Namespace):
+    """Loads a spreadsheet with caselaw URLs and downloads and parses each
+    decision.
+
+    Args:
+      :obj:`argparse.Namespace`: command line parameters namespace
+    """
+
+    n = 0
+    with output_stream(args.output) as fh:
+        csvout = csv.writer(fh, dialect="excel")
+        for uri in load_uris_from_csv(args.uris):
+            print(uri)
+            decision = Decision(uri=uri)
+            if args.download:
+                download_decision(decision, args)
+                csvout.writerow(decision.row)
+                n += 1
+                if args.limit and n >= args.limit:
+                    break
 
 
 def output_stream(outfile):
@@ -235,6 +269,25 @@ def download_decision(decision, args):
         fh.write(json.dumps(decision.values, indent=2))
 
 
+def load_uris_from_csv(csvfile):
+    """
+    Read a CSV file and get everything which looks like a caselaw URL and
+    return it as a Generator of str
+    Args:
+        csvfile(:obj:`pathlib.Path`)
+    Return:
+        Generator(str)
+    """
+    url_re = re.compile(CASELAW_BASE_URL + "(/decision/[a-f0-9]+)")
+    with open(csvfile, "r", newline="") as csvfile:
+        reader = csv.reader(csvfile)
+        for row in reader:
+            for val in row:
+                m = url_re.match(val)
+                if m:
+                    yield m.group(1)
+
+
 def main(args):
     """
     Args:
@@ -249,17 +302,20 @@ def main(args):
         if args.list:
             list_courts(args.list)
         else:
-            if not (args.courts or args.tribunals):
-                _logger.error(
-                    """
+            if args.uris:
+                download_uris(args)
+            else:
+                if not (args.courts or args.tribunals):
+                    _logger.error(
+                        """
 You must select at least one court or tribunal.
 
 Use the --list courts or --list tribunals options for a list of available
 options.
 """
-                )
-            else:
-                run_query(args)
+                    )
+                else:
+                    run_query(args)
 
 
 def run():