+++ CHANGELOG.rst
@@ -2,6 +2,22 @@
 Changelog
 =========
 
+Version 0.1.3
+=============
+
+Features:
+
+* Rate-limiting pause between requests can now be passed as a parameter to
+  the Search object
+
+* Search.url property which returns the URL which will be requested
+
+* command-line tool can download cases from a spreadsheet with caselaw URLs
+
+Bugs:
+
+* command-line tool was breaking when --download flag not on
+
 Version 0.1.2
 =============
 
@@ -13,8 +29,6 @@ Bugs:
 
 * fixed a bug in handling new-format cases which was breaking some Court of
   Appeals cases
-
-
 
 Version 0.1.1
 =============